feat: Add workflow for changed files (#16)
The changes in this PR will output the files that have changed, which
allows us to conditionally run the terraform actions if terraform files
have changed. If terraform files have not changed, we can skip
initialising terraform, because it does not need to be linted.

We can also use the changed files action to set the terraform working
directory.