attempt to run `required_status_checks` on main
When semantic-release pushes to main, the previous commit messages
included `[skip ci]` so `required_status_checks` stayed in a pending
state. This has been removed from the commit message so workflows are
triggered.

A generic workflow has been added for the `Validate PR title` workflow
to ensure this check also runs successfully on pushes to main.