test: required status checks on main for release workflow (#52)
Currently semantic-release runs on pushes to main to bump the release
version and update the CHANGELOG. It uses the 3ware-release app for
authentication - which is permitted to bypass pull requests. However,
due to the required status checks assigned to main branch's protection
rule, the app requires Administrator permissions to bypass the checks.

There are number of long running discussions requesting a *skip checks*
feature for this type of workflow:

https://github.com/community/community/discussions/13836

GitHub's documentation does provide a solution for **Handling Skipped
But Required Checks**:
https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/defining-the-mergeability-of-pull-requests/troubleshooting-required-status-checks#handling-skipped-but-required-checks

> You can fix this by creating a generic workflow, with the same name,
that will return true in any case similar to the workflow below

This pull request removes the `[skip ci]` syntax from semantic-releases
commit message.

Generic workflows for the pull request title validation and lint workflows have
also been added - as suggested in GitHub's docs.

Admin permissions will be removed from the application for testing the
new workflow.