feat(get-terraform-dir): Add second "get-terraform" step
The first step will output a JSON object which can be used to define a
matrix strategy for calling workflows.

The second step will output text that is used in the job summary.

Having two steps maybe inefficient, but converting the JSON into text,
particularly with multiple directories, was proving a challenge. Better
to keep it simple when the action will natively output the formats
required.