refactor(pr-title): Use environment variable pr error message
The conditional logic to dynamically generate the message for the pr
comment was too complex. This changes simplifies the logic by setting
the variable if the title is too long and using that, or if the title
is the correct length, using the error message from the semantic pr
action to set the value of the variable. The pr comment action will use
whatever the value is set to - no expression required.