fix(terraform-docs): Use input for tf-directory in all cases
There was a disconnect between the tf-docs step and the commit step
because the tf-docs step was using the input but the commit step had
'terraform' hardcoded. This is resulted in workflows reporting success
but docs were not updated because files were staged by tf-docs but not
located by the commit step.