fix(pr-title): Remove `if: always()` from success steps
if: always() is required to run steps that deal with failures. For
example, when the PR title is invalid, the step fails but the PR
comment should be created and the failure summary should be added.

Delete Comment and valid title summary will always run anyway because
the preceding steps have succeeded.